REFACTOR: Moved messages and tests of distribution module to own file.
