+++ pkg/indexer/decode/handle/test/group_test.go
@@ -13,6 +13,8 @@ import (
 	"testing"
 )
 
+// MsgVote
+
 func createMsgVote() types.Msg {
 	// Data from: 0A4BA0A30449C3269F313B5D974560F8D3A8179BE994054724898FF2D6866928
 	m := group.MsgVote{
